switch liquid class on pipette
We can now swap the liquid class with a new set_liquid_class_by_name
method that will change over the liquid class, as long as it wouldn't
invalidate the current settings _too_ badly. When we swap, we get the
new settings.

If you've overridden a flowrate or acceleration, the setting will be
kept; if it's the default, it will get updated to the default for the
new setting.