refactor: implement ConfigureForVolume
Unfortunately, the virtual pipette data provider now must be stateful
because it's emulating something stateful, and we should not just move
the state to state because it would be doubling for unavoidable hardware
state like the position of the pipette, and hardware controller really
wants to run this.

That said, we can now nicely use the exact same logic for the virtual
pipette mirror as the actual hardware control.

Needs new tests but doesn't break old ones.