refactor(api): Add PrivateResults to PE commands
When PE loaded pipettes, it would set the static config in its state by
having the load method of the equipment handler side-effect emit a
special action that only existed to update the configuration. This
worked ok as long as you weren't too picky about wanting to be sure that
the action was handled before the result of the command was evaluated.

The problem is that now we do care about that, a lot (though it's not
quite as evident in this commit). When we switch pickup and drop tip to
reset the tip state of a pipette, we're going to need to set the
configuration using this new private result method and rely on the
results in the command results handler.

This is implemented by poking a new typevar into CommandUpdateAction
that can optionally carry a "private result", something that will not be
exposed via HTTP or via the database but can carry information in a
lifetime-synchronous way to the result.

This required a lot of refactoring, but I think it's ended up pretty
cleanly; the only test changes are directly required by the changed code.