Refuse to get offset and tl on flex
Make the legacy routes for tip_length and pipette_offset 503 on flex and
not try and get the offsets using the old calibration accessors, since
they are hardcoded to use the ot2 versions of calibration which do not
apply to the flex and dumped errors in the logs all the time.

the api side change supports testing.