fix(app, app-shell): Support protocol backwards app compatibility (#13518)
Closes RQA-1509

* fix(app-shell): add protocol migration for v7.0.0

Downgrading the app from v7.0.0 to v6.3.1 or earlier causes whitescreening when protcols that use
new modules are present. This migration is a temporary shim to enable backwards compability until
OT-2 and Flex app parity is achieved.

Co-authored-by: Seth Foster <seth@opentrons.com>
Co-authored-by: Jethary Rader <66035149+jerader@users.noreply.github.com>

* feat(app): add unknown instrument/module error boundary

Render an error message instead of crashing if a user views a protocol with an instrument or module
that does not yet exist. This behavior is most likely triggered when a user downgrades their app.

Co-authored-by: Seth Foster <seth@opentrons.com>
Co-authored-by: Jethary Rader <66035149+jerader@users.noreply.github.com>