fix(app): Remove delete module calibration data
The module calibration overflow menu allowed you to delete a single
module's calibration data. This was (a) unimplemented (b) a bit of an
uphill battle to implement and (c) unlike all the other calibration
options in the calibration overview page, which just have the ability to
rerun calibration from that overflow. So remove the ability to delete
calibration data here.

Closes RQA-1526