feat(app,api): clear module calibration  (#13548)
Implements clearing module calibration on the flex and ties it into 
device reset, while removing the per-module info deletion that remains
unimplemented.

The module calibration overflow menu allowed you to delete a single
module's calibration data. This was (a) unimplemented (b) a bit of an
uphill battle to implement and (c) unlike all the other calibration
options in the calibration overview page, which just have the ability to
rerun calibration from that overflow. So remove the ability to delete
calibration data here.

Add clearing module calibration to the desktop and on-device display
device reset.

Closes RQA-1526
Closes RQA-1555
Closes RQA-1582