feat(robot-server): monitor door status (#13584)
Adds a route `GET /robot/door/status` paralleling the estop routes to tell you whether the door is open or closed, and whether the door is required to be closed to run a protocol. The required flag is always true on a flex and depends on the state of the door safety switch feature flag on the OT2, so clients can check its state to determine whether to add door-close UI rather than having to have machine-specific logic.