fix(app, components): fix image content in pipette, instrument and module cards
The styling of pipette, instrument, and module cards have diverged significantly and do not align
images and text with each other on the device details page. We would benefit from a larger
refactoring of these cards for consistency, but this PR introduces a temporary fix to align image
and text content on these cards.