fix(app,app-shell): undelete old robot update urls
We removed the urls for the old OT-2 infrastructure, because we don't
want them anymore. But because we removed them from the app config, and
the app doesn't downgrade configs, then if you
- ran 7.0, the config gets deleted
- run 6.3.1 or previous, the config never gets restored
and the app just doesn't know where to get robot system updates and
therefore acts as if there are none. That in turn means you can't
downgrade your robot through the app, you have to manually update from
zip.

The gross but only solution is to change the config back to get those
values to be present again when you downgrade. This leaves a historical
hole - someone who ran 7.0.0 and downgraded to 6.3.1 would still have
that behavior - but it's fixed going forward, and those affected can
just remove their app config.