chore(app): use cloudfront for release installs
We're properly pointing at the builds.opentrons.com _bucket_ for release
installs, but we were using the s3 provider config which will hit the s3
bucket's public url and not hit the CDN (or the DNS, for that matter,
which is like half the point of doing this...). So change that to look
at the actual normal url https://builds.opentrons.com/app in the same
way we do for internal releases.