fix(api): remove gratuitous tip state check prior to drop tip (#13450)
Before dropping tip in the hardware controller for ot2 and flex, we previously asserted the pipette
hc tip state is has_tip=False. This provides no extra security as that tip state could be stale. in
the worst case it prevents an intentional command to descend the tip ejector from occuring because
of an inaccurate sw state.

Co-authored-by: jbleon95 <jeremybenleon@gmail.com>
Co-authored-by: Max Marrone <max@opentrons.com>