fix(app,api,robot-server): expose flex serial (#13791)
The flex doesn't store its serial number in /var/serial; instead, it
stores it in an onboard eeprom. While we had the code to read it, we
weren't using that code anywhere, and we should expose it to the app for
display.

However, the update server doesn't (and shouldn't) know how to talk to
the onboard eeprom. That's a problem, because it knows perfectly well
how to read a file, and thus is the current source of truth for the
serial.

So this PR
- adds the serial to the robot server health endpoint, on both ot-2 and
flex
- changes the app to try and prefer the serial from the robot server
health endpoint, falling back to the update server health endpoint if
necessary