fix(robot-server): add default to robot serial
Some combination of pydantic and fastapi (fastapi, I'm pretty sure, due
to things like https://github.com/pydantic/pydantic/issues/1223 )
conflate "null" and "not present" in weird ways. This means that in the
health response pydantic class,
- If you create robot_serial in the proper way to have a pydantic field
that can be either a string or null but must always be passed in to the
constructor - annotation Optional[str] and do not provide a default -
then pydantic will say there's a field missing if you explicitly pass
in null (the bug this fixes, since this 500s the route)
- If you create robot_serial with a default value, which is _not
correct_ because there is no default, then something (probably fastapi)
will remove the _key_ from the model rather than issueing it with a
null, which is terrible, but is better than the other way I guess.

Updating fastapi might help with this maybe. I don't know.