fix(app): fix attachment flow firmware step filter (#13823)
Closes RQA-1827

* fix(app): fix attachment flow firmware step filter

Fixes the logic for determining whether or not the firmware step is necessary during an attachment
flow. Because it's not possible to check whether an instrument or module needs a firmware update
until it is attached, do not filter out the step entirely. Instead, if the instrument is known to
have the latest firmware, gracefully proceed to the next step.