make it prepare _to_ aspirate
The phrasing makes a lot more sense in context, so change the protocol
api name and the protocol engine command name to
PrepareToAspirate/prepare_to_aspirate rather than for. Do _not_ change
the hardware controller name because other stuff depends on that and I
personally like it more, and correspondingly do _not_ change the engine
pipetting handler name because that should correspond with the hardware interface.