fix(app): fix robot update via usb
After making the initial query to the robot server to create the robot update session, Robot updates
via USB utlize the top-level ok property of the response body to determine whether the robot-server
is ready to begin the update. The ok property is special cased for USB and expects a status text to
exist in the response body, however this status text does not match the usual expectation when
uploading a file to initiate the update. The ok property for USB is now determined by status code in
attition to status text.

Co-authored-by: Shlok Amin <shlokamin@me.com>
Co-authored-by: Brent Hagen <brent.hagen@opentrons.com>