app(feat): launch drop tip wizard logic (#13818)
Closes RAUT-795

* fix(app): relocate nice-modal provider to DesktopApp and OnDeviceDisplayApp

Certain hooks within modals require ApiHostProvider, which is context provided in
OnDeviceDisplayApp. Relocate modal rendering so hooks may utilize hostname.

* refactor(app): update PipetteData type instrumentModel property as PipetteModel type

* refactor(app): rename home-gantry icon to reset-position

Rename the home-gantry icon to reset-position, which is in-line with designs. Because this icon now
launches drop-tip flows, a more semantic name is appropriate.

* feat(app): wire up drop tip wizard entry from odd instrument detail overflow menu

Adds a new overflow menu on the instrument detail page with recalibrate and drop tip options.

* refactor(app): update ODD InstrumentDetail entry point given design discussion

* feat(app): add a drop tip banner

* feat(app): wire up protocolrunheader with drop tip wizard

* feat(app): determine if pipettes have tip attached after a protocol run

* feat(app): wire up odd runsummary page drop tip wizard entry point

* refactor(robot-server): check if protocol run is in terminal state instead of played

Drop tip flows introduce the need to create a maintenence run while a protocol run has played.
Instead of checking for the played status, check if a protocol run is in a terminal state.