don't try and use jsonschema for this
Jsonschema is a lot of things but one thing it isn't is a definer of
schema behavior based on values in the document. That's the key to all
of this - the document specifying a subschema for part of its content -
so we're just going to do it in code instead.

That means
- In the protocol schema, dynamically subschema'd objects get typed as
objects
- We get a new typescript validator to handle it all
- Also update python for this but it matters a bit less because the
python defines what is acceptable anyway, so there's no point trying to
be future proof here when downstream consumers will have to change