feat(shared-data): json protocol schema 8 (#13848)
Add JSON protocol schema 8, which is similar in semantic concept to json
protocol schema 7 but
- Requires user specification of the schema for the command List
This allows json protocols to switch which commands they're interning
without requiring further bumps to the protocol schema
- Does the same thing for command annotations, liquids, and labware
For the same reasons.

This should be a big step forward in flexibility for our protocol schemas.

Unfortunately, these new structures cannot be represented in jsonschema
because the schema to which components adhere is specified by the value
of another key in the document; jsonschema relations must be computable
at schema parse/verification time. We could solve this by using json hyper
schema but this seems a little much. Instead, there's a new js module for
parsing schema v8 that we'll introduce to the rest of the stack over time.
---------

Co-authored-by: Brian Cooper <brian@opentrons.com>