actually make integration tests fail if server doesnt start
Before, we'd wait infinitely for the server to start. If we made an
error in the server code such that it wouldn't come up, the only time
we'd notice is if we busted a timeout. So add a 30 second timeout to
starting the actual server that fails the tests rather than cancelling them.