fix(usb-bridge): dont reset input buffer on open
Serial will reset the input buffer when you open a new port, which is
bad if you are running a _listener_ and it's possible that the sender
might send data before you have the port open - which is definitely
possible with our usb gadget.

One way to fix this (there's no good way, it's not exposed as an option)
is to override the function that resets the input buffer in the classdef
- so new instances pick up the do-nothing override and don't reset the
buffer in open() - and then restore a bound-method version of the old one.