feat(app, app-shell): add download progress to robot update flows (#13915)
* feat(app): add download progress to robot update flows

If the robot update cache files are still downloading, no longer is an error thrown. Instead, the
update modal renders and a downloading stage of the update flow begins. Additionally, release files
are downloaded and returned to the browser layer separately from system files, so users may see
release notes as soon as they are downloaded.