chore(shared-data): pipette: refactor pickup, drop
These configs had a couple things to change:
- pickup and drop actually represent configuration for two different
actions: the press fit / plunger eject sequence that the low throughput
pipettes do, and the cam action sequence that the high throughput
pipettes do when handling a full rack. information was actually
overloaded and lost: there was only room for one speed for both separate
actions. By separating the configs for each action, we get that
information back
- fold in the per tip current information into the pickup config so it
looks a bit more natural and also can be distributed between the action kinds