chore(api): adapt to new tip configs
Now that we have separate configuration elements for picking up tips
with press fit vs cam action and dropping them with plunger eject and
cam action, we need to use them in a way that is nicely integrated with
the nozzle configuration stuff. That means that the plan methods should
check the current nozzle configuration and use the appropriate config
and behavior.