chore(robot-server): use partial tip pickup in cal
When we're calibrating an OT-2 multi channel, we do it with just one
tip. This means we have to drop the current to have the tip attached
with the appropriate amount of force. We now have a bunch of other code
focused on doing exactly that tied up with the partial tip pickup
mechanics, so use it instead.