fix(shared-data): Fix the pipette schemas, tests (#13949)
The tests that were supposed to ajv the pipette definitions weren't set
up properly, and the tests that were supposed to test that the tests
were set up properly weren't set up properly, so we weren't testing the
schemas so the schemas were wrong. Fix the schema-test setup, the
test-setup-schema-test setup, and the schema.