fix(api,shared-data): Use symbols for partial tips
We had implemented partial tip configurations by doing math on the
offsets of all the nozzles. This works, but has two downsides:
1. It's sort of hard to read unless you have a good mental model of the
physical layouts
2. It relies on doing exact numerical comparisons

In general, I think it is never a good idea to try and go _from_ numbers
_to_ symbolic knowledge. It's too vulnerable to small perturbations, and
it's too annoying to read.

Instead, let's encode the structure of the pipette nozzles in the
pipette geometry defs by writing down the rows and columns of the
pipette layout. Then, we can do our partial configuration logic based on
the rows and columns of the pipette.

We now do not hit a bug when we request a partial configuration that
doesn't include A1; and we now do not rely on array stride math; and we
now technically support arbitrary axis-aligned rectangles (at least on
this layer - definitely verboten above).