fix(app): do not require probe presence on 96 (#13986)
The 96 channel pipette is not capable of continuously checking for the
presence of a tip (or indeed a calibration probe). It needs to run a
special routing, exposed through ot3api.get_tip_presence_status() or
ot3api.verify_tip_presence(), to check for a tip. This happens
automatically in most of the places we use tips - aka inside other
protocol engine commands - but it does not happen automatically and all
the time.

This is usually fine, but when we're doing calibration we do it through
a protocol engine maintenance run where the way we interact with the
robot is dispatching PE commands - and we don't ever dispatch a PE
command that implicitly causes a tip check. That means that the 96 can't
be relied on to get the presence of a tip.

The long term fix for this is to add a checkTipPresence command to the
engine, at least as a maintenance command; in the short term, we can
disable the presence check for 96 channel pipettes to unblock testing.

Closes RQA-1892