fix(shared-data,api): Separate lt and ht pickup and dropoff (#13962)
These configs had a couple things to change:
- pickup and drop actually represent configuration for two different
actions: the press fit / plunger eject sequence that the low throughput
pipettes do, and the cam action sequence that the high throughput
pipettes do when handling a full rack. information was actually
overloaded and lost: there was only room for one speed for both separate
actions. By separating the configs for each action, we get that
information back
- fold in the per tip current information into the pickup config so it
looks a bit more natural and also can be distributed between the action kinds


Now that we have separate configuration elements for picking up tips
with press fit vs cam action and dropping them with plunger eject and
cam action, we need to use them in a way that is nicely integrated with
the nozzle configuration stuff. That means that the plan methods should
check the current nozzle configuration and use the appropriate config
and behavior.


When we're calibrating an OT-2 multi channel, we do it with just one
tip. This means we have to drop the current to have the tip attached
with the appropriate amount of force. We now have a bunch of other code
focused on doing exactly that tied up with the partial tip pickup
mechanics, so use it instead.

We had implemented partial tip configurations by doing math on the
offsets of all the nozzles. This works, but has two downsides:
1. It's sort of hard to read unless you have a good mental model of the
physical layouts
2. It relies on doing exact numerical comparisons

In general, I think it is never a good idea to try and go _from_ numbers
_to_ symbolic knowledge. It's too vulnerable to small perturbations, and
it's too annoying to read.

Instead, let's encode the structure of the pipette nozzles in the
pipette geometry defs by writing down the rows and columns of the
pipette layout. Then, we can do our partial configuration logic based on
the rows and columns of the pipette.

We now do not hit a bug when we request a partial configuration that
doesn't include A1; and we now do not rely on array stride math; and we
now technically support arbitrary axis-aligned rectangles (at least on
this layer - definitely verboten above).

---------

Co-authored-by: Laura Cox <31892318+Laura-Danielle@users.noreply.github.com>