fix(api): correct logic for simming partial tips
When we simulate partial tip pickup we do it by emulating the hardware
control logic in the virtual pipette data provider and then passing the
result to a nozzle manager the way the hardware controller does.
Unfortunately, the logic in the pipette virtual data handler did not
match the hardware controller.

Until we get to a place where we can have an actual simualting hardware
controller or something, we just have to keep these in sync.