feat(api): add support for dropping tips in the new trash bin and waste chute entities
Dropping tips previously required a reference well, passed in either explicility or implicitly. In
the implicit case, the protocol referenced a single-well "fixed trash" labware object that was
loaded automatically upon protocol instantiation. Here, I add/refine new classes for TrashBin and
WasteChute "disposal locations," add a protocol context method for loading each of these entities,
and add instrument support for dropping tips in these non-labware entities. The protocol core tracks
the presence of loaded TrashBins and WasteChutes to accommodate tip drops in the event that the user
did not specify a drop tip location or tie a trash_container to a their loaded pipette.