feat(app,shared-data): get addressable area for legacy fixed trash
provides a special case for the addressable area of the legacy Flex fixed trash found in older
protocol. parses the labware key of a protocol analysis and inserts addressable area moveableTrashA3
if a legacy fixed trash labware is found. moves the getSimplestDeckConfigForProtocol function to
shared-data and updates references in app and components.

closes RAUT-888