Update OT-2 fixed trashes.
I don't understand how to interpret the dimensions in the old fixed trash labware, and I'm scared of messing with the drop-tip point, so I guess let's make these 0-sized points in the same place as the old dropTipsOffsets.

This means the fixedTrash and shortFixedTrash addressable areas will no longer have the same origin as the v3 defs' slot 12. That's fine and actually makes sense. Those addressable areas are supposed to be equivalent to a v3 well position, not a v3 deck slot position.