fix(app-shell): use a wrapping stream for usb (#14214)
Implement the tcp socket emulator by creating a separate stream
that is piped through to the usb connection via its downward interface
rather than using the same stream for both the port and the socket or
pipelining the two streams together.

This implementation allows the lifecycles of the USB port connection,
which we want to be the same as the physical robot connection; and the
tcp socket emulators, which we want to be more like tcp sockets; to be
separate, which really increases reliability because we don't have the
port going up and down all the time anymore. This reduces spurious
disconnects.

It will also allow us to add socket activity timeouts to help with the
windows dropping data problem, though we can't really do that until the
http requests that cause synchronous actions that we make get keep alive
streaming responses.