fix(api): center only active column on 12-well
We have a system to center pipettes appropriately on the wells of
labware that span multiple "traditional" wells. For instance, in a
12-well reservoir, we need to move the center of an eight channel
pipette to the center of a given well of the 12-well reservoir.

However, that code was moving the center of _any_ pipette to the center
of a 12-well reservoir, inclunding the center of the 96-channel, which
is definitely not right. Instead, what we want to do is move the center
of the "active" column (the column that contains the active nozzle,
which is the nozzle that is moved when a caller specifies a NOZZLE
critical point) to the center of the well. And we only want to do this
on labware that have roughly one well per column.

The fix is in two parts.

First, a Y_CENTER critical point needs to exist. This critical point is
aligned in the X direction with the active nozzle of the pipette, but is
halfway between the Y positions of the backmost and frontmost nozzles in
the column containing the active nozzle of the currently active
nozzlemap - so, on an eight channel with a full map it would be the same
as XY_CENTER; on a single, an eight channel with a single nozzle map, or
a 96 with a single nozzle map, it will be the same as NOZZLE (and the
same as XY_CENTER); and on an eight channel or 96 with a partial-column map it will
be halfway up the partial.

On a 96 with a rectangular map (i.e. either full or quadrant) it will be
halfway between the back left and front left nozzles.

Next, we have to actually use that critical point. The engine would
check for a labware to have the centerMultiChannelOnWells quirk and emit
XY_CENTER for both source and destination if it was found; we now also
consider how many wells the labware has if it has
centerMultiChannelOnWells and (somewhat hackily, sadly) emit
- Y_CENTER, if the labware has the centerMultiChannelOnWells quirk and
1 < wells < 96
- XY_CENTER, if the labware has the centerMultiChannelOnWells quirk and
wells == 1, wells >= 96

This should probably change eventually into a new quirk called like
centerActiveColumnOnWells or something but that's for a different time.

Closes RQA-2118